fix: CalloutBlockComponents by renaming the file
